pyproject.toml
[project]
name = "project-runway"
version = "0.0.1"
description = "Add your description here"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE.md" }
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.16.2",
    "openai>=1.99.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
